@@ -1,36 +1,34 @@
 package ccrypto
 
 import (
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
 	"crypto/sha256"
-	"crypto/x509"
 	"encoding/base64"
-	"encoding/pem"
 	"fmt"
+	"os"
 
 	"golang.org/x/crypto/ssh"
 )
 
-//GenerateKey for use as an SSH private key
+// GenerateKey generates a PEM key
 func GenerateKey(seed string) ([]byte, error) {
-	r := rand.Reader
-	if seed != "" {
-		r = NewDetermRand([]byte(seed))
-	}
-	priv, err := ecdsa.GenerateKey(elliptic.P256(), r)
-	if err != nil {
-		return nil, err
-	}
-	b, err := x509.MarshalECPrivateKey(priv)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to marshal ECDSA private key: %v", err)
-	}
-	return pem.EncodeToMemory(&pem.Block{Type: "EC PRIVATE KEY", Bytes: b}), nil
+	return Seed2PEM(seed)
 }
 
-//FingerprintKey calculates the SHA256 hash of an SSH public key
+// GenerateKeyFile generates an ChiselKey
+func GenerateKeyFile(keyFilePath, seed string) error {
+	chiselKey, err := seed2ChiselKey(seed)
+	if err != nil {
+		return err
+	}
+
+	if keyFilePath == "-" {
+		fmt.Print(string(chiselKey))
+		return nil
+	}
+	return os.WriteFile(keyFilePath, chiselKey, 0600)
+}
+
+// FingerprintKey calculates the SHA256 hash of an SSH public key
 func FingerprintKey(k ssh.PublicKey) string {
 	bytes := sha256.Sum256(k.Marshal())
 	return base64.StdEncoding.EncodeToString(bytes[:])
