@@ -4,6 +4,6 @@ package chshare
 //incompatible changes are made, this will
 //be incremented to signify a protocol
 //mismatch.
-const ProtocolVersion = "chisel-v3"
+var ProtocolVersion = "chisel-v3"
 
 var BuildVersion = "0.0.0-src"
